backend/main.py
import os
import re
from typing import List, Optional
from fastapi import FastAPI, HTTPException, Request
from fastapi.responses import HTMLResponse, JSONResponse
from fastapi.staticfiles import StaticFiles
from fastapi.middleware.cors import CORSMiddleware
from pydantic import BaseModel
from dotenv import load_dotenv
from groq import Groq

# Load environment variables
load_dotenv()

# Initialize Groq client
client = Groq(
    api_key=os.environ.get("GROQ_API_KEY"),
)

app = FastAPI()

# Enable CORS
app.add_middleware(
    CORSMiddleware,
    allow_origins=["*"],
    allow_credentials=True,
    allow_methods=["*"],
    allow_headers=["*"],
)

# Serve static files for frontend
app.mount("/static", StaticFiles(directory="../frontend"), name="static")

# Models
class ReviewRequest(BaseModel):
    code: str
    language: str
    focus_areas: List[str]

class RewriteRequest(BaseModel):
    code: str
    language: str
    focus_areas: List[str] # Added to pass focus areas if needed for rewrite context

# --- API Endpoints ---

@app.get("/", response_class=HTMLResponse)
async def read_root():
    with open("../frontend/login.html", "r") as f:
        return f.read()

@app.get("/app", response_class=HTMLResponse)
async def read_app():
    with open("../frontend/index.html", "r") as f:
        return f.read()

@app.post("/api/review")
async def review_code(request: ReviewRequest):
    try:
        prompt = f"""You are a senior software engineer with 15+ years of experience.
Analyze the following {request.language} code focusing on: {', '.join(request.focus_areas)}.

Provide output in EXACT structure:

🔴 Critical Issues

bullet points

🟠 High Priority

bullet points

🟡 Medium Priority

bullet points

🟢 Low Priority

bullet points

📌 Overall Summary

Short summary paragraph.

Code:

{request.code}
"""
        completion = client.chat.completions.create(
            messages=[
                {
                    "role": "user",
                    "content": prompt,
                }
            ],
            model="llama-3.3-70b-versatile",
            temperature=0.3,
            max_tokens=2000,
            top_p=0.9,
        )
        
        review_text = completion.choices[0].message.content
        parsed_review = parse_review_response(review_text)
        parsed_review["raw_review"] = review_text
        
        return JSONResponse(content=parsed_review)

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

@app.post("/api/rewrite")
async def rewrite_code(request: RewriteRequest):
    try:
        prompt = f"""You are an expert software architect.
Rewrite the following {request.language} code to:

- Fix all bugs
- Improve performance
- Remove security vulnerabilities
- Apply best practices
- Add docstrings/comments
- Make it production-ready

Provide:
1. Rewritten code only
2. List of improvements

Code:
{request.code}
"""
        completion = client.chat.completions.create(
            messages=[
                {
                    "role": "user",
                    "content": prompt,
                }
            ],
            model="llama-3.3-70b-versatile",
            temperature=0.3,
            max_tokens=2000,
            top_p=0.9,
        )
        
        response_text = completion.choices[0].message.content
        
        # Simple extraction logic (can be refined)
        # Assuming the model follows instructions, we might need to parse
        # But for now, returning the raw text which the frontend can render is safer 
        # unless we strictly enforce JSON output from LLM, which is harder with just text prompting.
        # Let's try to structure it a bit if possible, or just return as is.
        # The user requirement say "Return { rewritten_code: "", improvements: [] }"
        # We'll try to parse typical markdown code blocks.
        
        code_match = re.search(r"```(?:\w+)?\n(.*?)```", response_text, re.DOTALL)
        rewritten_code = code_match.group(1) if code_match else response_text
        
        # Extract improvements (assuming they are listed after the code or before, usually formatted as list)
        # We can just return the full text for specific sections if strictly parsing is too brittle without JSON mode.
        # However, let's try to provide the requested structure.
        
        return JSONResponse(content={
            "rewritten_code": rewritten_code,
            "improvements": response_text # sending full text for now so frontend can display the list and everything
        })

    except Exception as e:
        raise HTTPException(status_code=500, detail=str(e))

def parse_review_response(review_text: str):
    """
    Parses the review text into structured sections.
    """
    sections = {
        "critical": [],
        "high": [],
        "medium": [],
        "low": [],
        "summary": ""
    }
    
    # Regex patterns for sections
    critical_pattern = re.search(r"🔴 Critical Issues(.*?)(?=🟠|🟡|🟢|📌|$)", review_text, re.DOTALL)
    high_pattern = re.search(r"🟠 High Priority(.*?)(?=🟡|🟢|📌|$)", review_text, re.DOTALL)
    medium_pattern = re.search(r"🟡 Medium Priority(.*?)(?=🟢|📌|$)", review_text, re.DOTALL)
    low_pattern = re.search(r"🟢 Low Priority(.*?)(?=📌|$)", review_text, re.DOTALL)
    summary_pattern = re.search(r"📌 Overall Summary(.*?)$", review_text, re.DOTALL)

    def extract_bullets(text):
        if not text:
            return []
        # Extract lines starting with hyphens, asterisks, or numbers
        return [line.strip().lstrip("-*•").strip() for line in text.strip().split('\n') if line.strip().startswith(("-", "*", "•"))]

    if critical_pattern:
        sections["critical"] = extract_bullets(critical_pattern.group(1))
    
    if high_pattern:
        sections["high"] = extract_bullets(high_pattern.group(1))

    if medium_pattern:
        sections["medium"] = extract_bullets(medium_pattern.group(1))

    if low_pattern:
        sections["low"] = extract_bullets(low_pattern.group(1))

    if summary_pattern:
        sections["summary"] = summary_pattern.group(1).strip()
        
    return sections

if __name__ == "__main__":
    import uvicorn
    uvicorn.run(app, host="127.0.0.1", port=8000)
